Change list:   The systems::Wam::moveTo() member function can now be used to simultaneously control tool position and tool orientation using pose_type   Updated examples/ex02_hold_position.cpp to demonstrate systems::Wam::moveTo(pose_type)   Added a template specialization for math::Spline that can interpolate boost::tuple types   API CHANGE: Removed the initialDirection parameter from the math::Spline constructor   Added Puck::saveAllProperties() member function   The sandbox/ folder is installed to ~/libbarrett_sandbox by default   Python bindings for libbarrett are now compiled by default (use "cmake . -DWITH_PYTHON=OFF" to disable)   Fixed a bug in the log::RealTimeWriter constructor

+++ include/barrett/log/detail/real_time_writer-inl.h
@@ -57,7 +57,7 @@ RealTimeWriter<T, Traits>::RealTimeWriter(const char* fileName, double recordPer
 	size_t recordsInSingleBuffer = 16384 / this->recordLength;
 
 	// with a factor of safety of 5, how many seconds to fill a single buffer?
-	period = static_cast<size_t>( (recordsInSingleBuffer * recordPeriod_s) / 5.0 );
+	period = (recordsInSingleBuffer * recordPeriod_s) / 5.0;
 	if (period < 0.003) {
 		throw(std::logic_error("(log::RealTimeWriter::RealTimeWriter()): This constructor was not designed for data rates this high."));
 	}
